QDoc: Initialize static members of `Sections` in-place
The `Sections` class has a series of static vector of `Section`
whose elements are initialized in two parts.

First, each vector is initialized the expected number of elements, only
setting certain properties of the newly created `Section`s, then, each
element in each of the vector is properly initialized when a `Sections`
instance is constructed.

A boolean flag is used to avoid performing this computation more than
once.

Furthermore, a series of late initialization methods are provided by
`Section` to support this workflow.

The properties of those static elements that are set during this
initialization process are not expected to change after the first
it is performed once.

To reduce the amount of cruft required to support his initialization, we
move the initialization to be performed in-place, once.

The various `init` overloads from `Section` were removed, as they were
only required to late-initialize some the relevant properties.

Similarly, the two constructors for `Section` were removed, as they only
partially initialized some of the relevant state, delegating further
initialization to the `init` methods.

To replace those methods, a single constructor was provided,
initializing all statically relevant properties at once.
The only direct usages of a `Section` constructor, in
`HtmlGenerator::generateDetailedMember` and
`DocBookGenerator::generateDetailedMember`, were modified to adhere to
the new constructor.

Some of the members of `Section` were reordered, to respect the ordering
of the new constructor, whose ordering is based on the most used
overload of `init`, whose original ordering was inconsistent between
overloads.

`Sections::initSections`, which was used during construction of a
`Sections` instance to perform the initialization of the static vectors,
was removed as unnecessary.

Its body, initializing the various elements of the various static
vectors, was moved to the top level of the file and modified to be an
in-place initialization for each static vector.

Change-Id: I2cda7a9d8c7fc3694ed5913a62b7dfa84c797075
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>