qdoc: Ensure links to navigation items are based on the correct target
The navigation bar (breadcrumbs) was composed of Atom instances of the
type Atom::NavLink. These links were resolved based on the name() of
the Node being linked to, which caused a potential for incorrect
linking if there were multiple nodes with the same name - case in
point,

  \module QtMultimedia
  \qmlmodule QtMultimedia

Class Atom operates solely on strings as link targets, but it does
offer Atom::LinkNode type which carries a Node pointer converted to
a string. Use this to ensure the links go to the correct target.

Add comments explaining the logic for adding links to module pages.

Fixes: QTBUG-106209
Change-Id: I2a327fcb3bf5a71596439fcd7449c275e31c0776
Reviewed-by: Luca Di Sera <luca.disera@qt.io>