QDoc: Remove the second parameter of Section::sortName
`Section::sortName`, an internal method that specifies an ordering over
`Node`s that is used to sort the incoming `Node`s into the category maps
used by a `Section`, accepts two parameters.

The first parameter is a `Node` pointer, from which a sortable `QString`
is extracted starting with the node's name.

The second parameter is a `const QString` (unnecessarily a pointer),
that is used as the base of the sortable `QString` instead of the node's
name if it is passed in.

The only usage of the second parameter is in `Section::add`, where
`sortName` is called referencing a `QString` that is the name of the
node passed to `sortName`.

That is, the only usage of the parameter results in the same behavior of
a unary call.

Hence, the second parameter is removed from `SortName`, as it is
unnecessary.

The declaration of `sortName` in the header file was modified to adhere
to the new interface.

The definition of `sortName` in the cpp file was modifierd to adhere to
the new interface and the code branching on the parameter was removed
from it.

The single call site of `sortName/2` was modified to avoid passing a
second argument.

The documentation for `sortName` was modified to remove any reference to
the second parameter.

Change-Id: I4d145eacd865b560a63cdde6beafafdc959bb184
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>