QDoc: Move `sortName` out of `Section`
`Section::sortName` is an internal method that is used to generate
sortable identifiers for `Node`s that are partitioned by a `Section`.

The method is currently private under `Section` but does not use any
internal state of `Section`, simply requiring any `Node` input.

Hence, the method was removed from `Section`'s interface in favor of
being a free-function visible only in the implementation file.

Change-Id: Ifa1be8a081edb2d31ddbeb7bcd41205191a18d51
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>