qdoc: Disable linking inside table-of-contents items
QDoc allows links inside section titles:

  \section1 \l QString usage

When generating the TOC, QDoc dumped the HTML generated from the list
of Atoms directly into the contents of the list item, including the
anchor tags.

While doing so, it marked the internal state of the generator to be
'in link'; this was not enough however, as a LinkAtom inside the
section title closed the actual (TOC) in-page link and started another
with its own link target. This resulted in confusing page TOC where
an item could link to somewhere else than *this* page.

Fixes: QTBUG-107602
Pick-to: 6.4
Change-Id: I5912e71b2ea1a1e5f04ef82012b7e105e4fc9da5
Reviewed-by: Paul Wicking <paul.wicking@qt.io>