QDoc: Fix ordering of overloads for QML Types
When QDoc generates a page for a QML type, as a result of the "\qmltype"
topic command, it generates, similarly to C++, a summary list of
members, a detailed list of members and a separate page listing a
summary of all members of the type.

Internally, QDoc uses two coupled classes, `Section` and `Sections`, to
destructure an `Aggregate`, the internal representation of documentable
element that are composed of multiple documentable elements, such as a
QML Type, into a series of ordered categories of child elements.

Those categories are later used to generate the various "sections" of
the `Aggregate` documentation, such as the various list, as described
above, for a QML Type.

The ordering of all those "sections" is given based on the results of
the free-function `sortName` in "sections.cpp".
`sortName` produces ordered strings based on a `Node`'s name.

When QDoc parses QML methods that are overloads, as it does for C++, it
assigns an "overload number" to each of them, to discern the various
equally named elements.

One of the elements is considered the "primary overload", to which no
number is assigned, and all other elements are considered "non-primary
overloads", to which an increasing positive integer "overload number" is
assigned.

Due to the current implementation of `sortName`, the "overload number"
is not considered when sorting the methods of a QML Type, contrary to
the behavior reserved for C++ overloads.

As a consequence, the various overloads in a section are ordered based
on the implementation details of their storing mechanism and sorting
algorithm.

In particular, in our specific case where a red-black tree map is used,
they are ordered in descending order of their overload number.

For example, given some method named X and a few of its overload X-1,
X-2 and X-3, the "sections" listing the overloads will order them as
x-3, X-2, X-1, X.

While the assignment of the "overload number", currently, generally only
follows the order in which QDoc encounters the various overloads, such
that it may not be consistent between source files or versions of the
code, it generally follows, in the Qt codebase, a "better looking"
order, where overloads with a "more complex" signature appear after
their peers with a "simpler" signature.

Furthermore, we enforce an ascending order for C++ overloads based on
the overload number, contrary to the current behavior for QML overloads.

Hence, to increase the consistency between the ordering of overloads in
C++ and QML documentation and possibly provide a "better looking"
output, `sortName` was modified to take into account the "overload
number" for QML overloads.

The regression files for `tst_generatedOutput` were regenerated to take
into account the ordering change.

Change-Id: I904ec35e428e8b0b1740826eb6ad9726a261938a
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>