qtattributionsscanner: Read license files from LICENSES directory
If the 'LicenseFile[s]' property is empty in the qt_attributions.json
file, we now extract the SPDX license identifiers from the SPDX license
expression in the LicenseId property.

The extracted SPDX license identifiers are used to locate license files
in the LICENSES directory up in the directory hierarchy.  If we
encounter a license ID without matching license file, we print an error
message.

This enables us to deduplicate license files in our repositories.

Pick-to: 6.4
Fixes: QTBUG-104126
Change-Id: I38b281c97e039a8158e143ffa16ba1966713d030
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>