lupdate: Explain -locations options in command help
The difference between "-locations absolute" and "-locations relative" is
only in the stored line information. For relative, the line number is
relative to a previous potential entry in the same file and prefixed
by a '+'. For absolute, the line number is just counting from the
beginning of the file.

Anyhow, from the existing help output this is unclear, and users might
expect that 'absolute' and 'relative' first and foremost applies to the
file path.

Fixes: QTBUG-107107
Change-Id: Ibc62df867b9fc914cf8c8fa076a555b35c240a54
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit 883a8ea1220c032586f2781e1503c2140870ffb0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>