qdoc: Stop listing invalid group names in type reference
QDoc was listing links to groups that a C++ class or QML type
documentation topic was part of - that is, all group names that appeared
as arguments to \ingroup command(s). However, not all group names used in
Qt docs have a matching \group topic. This resulted in listing of groups
that do not have a title and/or are not valid as link targets.

Filter out the list of group names into CollectionNode pointers for
groups that have been 'seen', which means that they do generate a group
page and can be linked to.

Fixes: QTBUG-108100
Change-Id: I2924d49d7bc9947fd4ac12c6f1c0daed96d52da6
Reviewed-by: Luca Di Sera <luca.disera@qt.io>