qdoc: Query system include paths from the compiler on Linux
In certain situations depending on the build environment and
the libclang version that QDoc links against, Clang may be
unable to resolve the correct system include paths
automatically. This seems to be the case particularly for the
portable QDoc binary that is provisioned into test builds
in the CI system, and results in spurious documentation
warnings.

Query the paths from GNU C++ compiler on Linux, just like
we already do for macOS/clang++.

Pick-to: 6.4
Task-number: QTBUG-108353
Change-Id: I92ffb4da0ce8af166084eb20a29ec4b00c047489
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>