QDoc: Order C++ copy/move constructors in class pages and members list
When QDoc generates the page for a C++ class, it generates a summary
list of the direct (non-inherited) members of that class (methods,
slots, ...) and a detailed list of those same members.

Furthermore, it generates an additional page that lists all members,
inherited or not, for that class.

Elements of those variouso lists are ordered under a certain
methodology.
For example, currently, QDoc orders the constructors at the top, before
other specially-ordered members.

Internally, QDoc has some special handling for copy/move constructors.
Indeed, a `FunctionNode`, the internal representation that QDoc uses for
C++ functions, has a "metaness" attribute that will be set depending on
the kind of element the `FunctionNode` represents, with special cases
for copy and move constructors.

Currently, QDoc orders all constructors "at the same levels",
independently of the kind of constructors they may be.
Due to an implementation detail, this will usually, but not necessarily
always, mean that the constructors will appear in the order they were
declared in the code.

That is, if some class X has constructors X1, X2, X3, and they are
declared in whatever header file in the order X1, X2, X3, they will be
ordered as X1, X2, X3 in the relevant generated lists.
If the same constructors where declared with order X2, X1, X3, in
whatever header file, they will be ordered as X2, X1, X3 in the relevant
generated lists.

The major implementation detail that provide this ordering, an internal
number that is assigned to overloads, is independent of the ordering
algorithm that QDoc uses.
Furthermore, the ordering that QDoc uses is not total, such that
distinct elements may not be uniquely orderable.

Those two properties together ties some of the ordering to the internal
implementation that the ordering routine uses, with the issue that
changing the implementation might change the ordering itself.

In particular, the current implementation is provided by the
free-function `sortName` in "sections.cpp", that is used by the
`Section`/`Sections` pair of classes to provide an ordered view of
certain lists that QDoc generates.

Internally, the result of a call to `sortName`, is used as the key in
one or more multimaps that later provide the required ordering, which is
implementation dependent when collisions are met.

A certain case where collision might happen has been between copy/move
constructors and other constructors, generally the default one, where
the same `sortName` is produced.
The collision behavior for those constructors might actually stem from a
different bug in the codebase that has not yet been investigated and
might later be addressed.

It is expected that the internal implementation that provides the
sorting might be changed in the near future and then changed again at a
later phase, such that those implementation dependent ordering might
modify the current look of the documentation.

To avoid the issue of the ordering changing between implementations,
`sortName` now guarantees that the result that is produced for each of a
constructor/copy constructor/move constructor is distinct from the one
produced for the other two kinds.

`sortName` uses prefix letters prefix letters prepended to the name of a
certain element to provide the main sorting. To implement the new
behavior, different prefix letters will now be applied to different kind
of constructors.

In particular, a "D" for a copy constructor, an "E" for a move
constructor and a "C" for each other constructor.

Due to the addition of new elements, to preserve the original ordering,
the prefix letter for other elements has now been changed. For example,
a destructor would previously have a prefix letter of "D" while now it
would have a prefix letter of "F".

The provided ordering now ensures that all non-copy, non-move
constructors will be listed, later followed by copy constructors and
finally followed by move constructors.

The ordering of non-constructor elements is unchanged.

A small chunk of code in `sortName`, dealing the ordering of classes,
was moved up in the function.
This result in no semantical changes but is expected to simplify the
top-down reading of the function, as it more closely mimics the final
ordering of the elements in the generated lists.

This change impacts the current documentation by moving around the
order of some copy/move constructors which were previously interleaved
with other constructors. The content of the documentation is not
expected to be touched by this change.

Change-Id: I1ee3e63d971644fbcd0b20a8ed1f55dcb0b30655
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>