qdoc: Avoid C++ parsing issues when linking against libclang 15
libclang from LLVM version 15 causes a lot of parsing failures with
the 'error code: 4' specifically when initialized with C++20 support.
This problem is highly likely to be fixed in LLVM 16 release.

Drop down to C++17 standard if libclang from LLVM 15 is detected.

There are a few places in the Qt API where certain features are
conditionally enabled for C++20 in a way that's also visible in
the generated documentation. These can be worked around separately.

Pick-to: 6.4
Fixes: QTBUG-94365
Change-Id: I5a3fd14dfe07102595fc941b27c46688b0f29697
Reviewed-by: Luca Di Sera <luca.disera@qt.io>