QDoc: Remove documentAll from Sections::buildStdCppClassRefPageSections
QDoc uses a pair of classes, `Section`/`Sections`, when generating the
final documentation, to organize and classify the known documentable
elements into categories that are later used to generate certain ordered
lists of items, such as the list of members that a C++ class has.

Internally, `Sections` dispatches an `Aggregate`, the internal
representation used for documentable elements that are composed of
multiple documentable elements, and its children into a series of
`Section` to categorize them.

Multiple methods are implemented to perform this dispatch, based on the
"kind" of aggregate that is being dispatched.

When `Sections` dispatches representations for C++ classes, structs of
unions, it performs the initial dispatch of the `Aggregate` through
`Sections::buildStdCppClassRefPageSections`, whose categorization is
later used, generally, during the generation of the lists in the
reference page of that same C++ element in the output documentation.

`Sections::buildStdCppClassRefPageSections` iterates, between other
things, the documentable elements that compose the `Aggregate`, and
their children if any, to categorize them.

A flag variable, `documentAll`, would branch the code based on the main
`Aggregate` that was being categorized.

When `documentAll` was false, any element that did not have a
user-provided documentation would be left out, thus not appearing in the
lists that would be generated by through the `Sections`.

When `documentAll` was true, this filter would be bypassed, generally
document all elements and their children independently of the presence
of any user-generated documentation.

`documentAll` always defaulted to true. It then was considered false if
and only if the dispatched `Aggregate` had a parent but didn't have a
name or a documentation.

It is expected by QDoc, but not enforced, that all `Aggregate`s have a
name, such that no `Aggregate` can trigger, unless a bug is present, the
condition to make `documentAll` false.

The only exception to this is the "root namespace", a "special"
`Aggregate` that sits at the top of the internal tree-representation of
documentable elements built by QDoc.
Nonetheless, the "root namespace" does not have a parent, such that it
couldn't trigger the condition either.

Furthermore, the "root namespace" has a "namespace" "kind", and is thus
dispatched to a different method, `Sections::buildStdRefPageSections`,
such that it would not pass by the same `documentAll` condition at all.

Hence, it should be theoretically impossible for the condition to
trigger, making the `documentAll` flag and branching superfluous, as it
is always expected to be true.

Hence, `documentAll` and its related branching was removed from
`Sections::buildStdCppClassRefPageSections` as dead code, to simplify
the method and remove part of its cruft.

The original reason for the presence of `documentAll` is unknown.
It is supposed that there was originally no distinction between methods
for the dispatch of differently-kinded `Aggregates`, such that the
condition was supposed to specifically trigger for certain elements and
lost its purpose during a refactoring of `Sections`/`Section`.

Nonetheless, no internal representation of the recent and less
recent times is known to sensibly be identified by the presented
condition, such that the condition might have been incorrectly
transcribed, might have never been correct to begin with or might have
depended upon certain bugs or behavior that have faded with time.

As the original reasoning for the condition is uncertain, it is possible
that certain now unknown bugs might resurface in the future, such that a
certain eye to unexpected changes in the output documentation should be
kept in the short-term.

This patch has no visible effect on the current the current
documentation output.

Change-Id: Iaefa5c04aa1bdf4e46f7f50cce03033f024d7ce8
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>