QDoc: Remove unnecessary branching
`HtmlGenerator::generateAllQmlMembersFile` would loop a vector
structure to generate the list of members for a Qml type.

The structure was traversed trough a loop bound on the size of the
structure, such that if the structure is empty no iteration would be
performed.

Nonetheless, an additional scope wrapped the loop block, avoid the
execution when the structure was empty, that is, when its size was 0.

The unnecessary additional block and indentation further cluttered the
code and was thus removed with no change in behavior.

Change-Id: I910cb13dfa2a1b91e7701e1c78c7f1808ed0f1b5
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>